feat(window.lua): allocate window using column list
When opening new window, it is required to specify an existing window
and a direction to split.

Allocating the windows based on the column list.

  1. Iterate through the first element of column lists and vsplit right.
  These first elements will be the top-window of each column.
  2. Iterate through other elements in each column and split below.

This allocation rule makes it easier to change window layout.

User can change the window layout through user config.

After allocation, "git_diff" and "git_show" windows are enlarged to
improve visibility.

Signed-off-by: asana17 <kei17.kaight084@gmail.com>